contrib: gunyah-hypervisor update 12/2023
C runtime updates:
* MISRA and Cyclomatic complexity updates
* Code coverage support
* Support debug build quality configuration

+++ arch/armv8/include/arch_reloc.h
@@ -7,6 +7,6 @@
 #define R_AARCH64_RELATIVE 1027
 
 #define ARCH_CAN_PATCH(r_info)                                                 \
-	(((R_TYPE(r_info) == R_AARCH64_NONE) ||                                \
-	  (R_TYPE(r_info) == R_AARCH64_RELATIVE)) &&                           \
-	 (R_SYM(r_info) == 0))
+	(((R_TYPE(r_info) == (uint32_t)R_AARCH64_NONE) ||                      \
+	  (R_TYPE(r_info) == (uint32_t)R_AARCH64_RELATIVE)) &&                 \
+	 (R_SYM(r_info) == 0U))
